move datasets/ and viewport/ into their own routers
